Title-1/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1
geopy==2.4.1
